// frontends/main/src/app/core/model/field-dto.ts
/**
 * Rodano API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WorkflowDTO } from './workflow-dto';
import { FieldModelDTO } from './field-model-dto';
import { PossibleValueDTO } from './possible-value-dto';
import { WorkflowStatusDTO } from './workflow-status-dto';


export interface FieldDTO { 
    /**
     * Scope reference
     */
    scopePk: number;
    scopeId: string;
    scopeCodeAndShortname: string;
    /**
     * Event reference
     */
    eventPk?: number;
    eventId?: string;
    eventShortname?: string;
    /**
     * Dataset reference
     */
    datasetPk: number;
    datasetId: string;
    datasetModelId: string;
    /**
     * Field pk
     */
    pk: number;
    creationTime: Date;
    lastUpdateTime: Date;
    /**
     * Field model
     */
    model: FieldModelDTO;
    /**
     * Field model id
     */
    modelId: string;
    /**
     * Is the field attached to a removed entity?
     */
    inRemoved: boolean;
    /**
     * Is the field attached to a locked entity?
     */
    inLocked: boolean;
    /**
     * Possible values of the field, if the field type is a multiple
     */
    possibleValues: Array<PossibleValueDTO>;
    value?: string;
    valueLabel?: string;
    newContent?: boolean;
    filePk?: number;
    fileName?: string;
    workflowStatuses: Array<WorkflowStatusDTO>;
    possibleWorkflows: Array<WorkflowDTO>;
}