// frontends/main/src/app/core/model/role-dto.ts
/**
 * Rodano API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ScopeMiniDTO } from './scope-mini-dto';
import { RoleStatus } from './role-status';
import { ProfileDTO } from './profile-dto';


export interface RoleDTO { 
    scopePk: number;
    profileId: string;
    pk: number;
    userPk?: number;
    robotPk?: number;
    profile: ProfileDTO;
    scope: ScopeMiniDTO;
    status: RoleStatus;
    enabled: boolean;
    canEnable: boolean;
    canDisable: boolean;
}
export namespace RoleDTO {
}